pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fewshoter"
version = "1.0.0"
description = "CLIP-based few-shot image classification toolkit"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [{ name = "Fewshoter Maintainers" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "torch>=1.9.0",
  "torchvision>=0.10.0",
  "open_clip_torch>=2.0.2",
  "opencv-python>=4.5.0",
  "Pillow>=8.0.0",
  "numpy<=1.26.1",
  "scipy>=1.7.0",
  "scikit-learn>=1.0.0",
  "matplotlib>=3.4.0",
  "seaborn>=0.11.0",
  "pandas>=1.3.0",
  "tqdm>=4.62.0",
  "PyYAML>=5.4.0",
  "imageio>=2.9.0",
  "Flask>=2.2.0",
  "Werkzeug>=2.2.0",
  "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "ruff>=0.6.0",
  "black>=24.0.0",
  "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/TianWen580/fewshoter"
Repository = "https://github.com/TianWen580/fewshoter"
Issues = "https://github.com/TianWen580/fewshoter/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.setuptools]
package-dir = { fewshoter = "." }

[tool.setuptools.packages.find]
where = ["."]
include = ["fewshoter*"]
